Enabled static linking of VanDerPol and SynchronousSupervisoryControl examples.
