+++ Clocks/model.c
@@ -3,6 +3,10 @@
 #include <math.h>
 #include <stdlib.h>
 
+
+void logError(ModelInstance *comp, const char *message, ...);
+
+
 /*
 
 time        0 1 2 3 4 5 6 7 8 9
