+++ LinearTransform/config.h
@@ -13,6 +13,7 @@
 #define MODEL_EXCHANGE
 
 #define SET_FLOAT64
+#define GET_FLOAT64
 #define GET_INT32
 #define SET_INT32
 #define EVENT_UPDATE
