+++ LinearTransform/model.c
@@ -1,6 +1,7 @@
 #include "config.h"
 #include "model.h"
 
+void logError(ModelInstance *comp, const char *message, ...);
 
 void setStartValues(ModelInstance *comp) {
